+++ public/data/StudentFeedbackData.json
@@ -1,0 +1,62 @@
+[
+    {
+        "id": 1,
+        "name": "North Central Province Career Compass Session - 2025",
+        "grade": "10,11",
+        "school": "@Rajarata university of Sri Lanka",
+        "date": "May 2025",
+        "province": "North Central Province",
+        "program": "CAREER COMPASS",
+        "comment": "ඇත්තටම මේ වැඩසටහන ගොඩක් හොඳයි! ජීවිතයට ගොඩක් වැදගත්, අපිට යන්න ඕනේ ඉලක්කය, පාර කියලා  දුන්නා.  මම ගොඩක් ආසයි  Maths කරන්න. ඉතින් ඒකට යන්න ඕනෙ පාර තේරුණා. මේ වැඩසටහන කරපු අයියලා අක්කලා ඔයාලට ගොඩක් ස්තුතියි."
+    },
+    {
+        "id": 2,
+        "name": "North Central Province Career Compass Session - 2025",
+        "grade": "10,11",
+        "school": "@Rajarata university of Sri Lanka",
+        "date": "May 2025",
+        "province": "North Central Province",
+        "program": "CAREER COMPASS",
+        "comment": "අද දවසේ අපි මේ විශ්වවිද්‍යාලයට ආව විදිහට නෙවෙයි එළියට යන්නේ. ඊට වැඩිය ටිකක් හරි වෙනස් වෙච්ච කෙනෙක් විදිහට. අද දවසේ අපේ ඉදිරි අනාගතය හදාගන්න උදව්කරා. ඊටත් වැඩිය කියුවොත් අපේ ජීවිතය ඉදිරියට ගෙන යන්න පුළුවන් විදියේ සැලැස්මක් හදලා දුන්නට ඔයාලට thank you . ඔයාලා අපිට හොඳ සැලසුමක් හදල දුන්නා වාගේ ඔයාලගේ ජීවිතය ද සාර්ථක වෙන්න කියලා ප්‍රර්ථනා කරනවා."
+    },
+    {
+        "id": 3,
+        "name": "Career Compass – Career Guidance Session for the Northern Province",
+        "grade": "10,11",
+        "school": "@TLH-1, Faculty of Technological Studies, University of Vavuniya",
+        "date": "25th June 2025",
+        "province": "North Province",
+        "program": "CAREER COMPASS",
+        "comment": "I understood about the higher education options, types of government institutions, types of job paths, working industries, free courses and external courses of Government, private universities, HND courses, NVQ courses, computer training, and SLITHM course, and if I fail what can I do also. And I got more knowledge about universities and the study system and future lifestyle and passion. So we are very thankful for this event."
+    },
+    {
+        "id": 4,
+        "name": "Career Compass – Career Guidance Session for the Northern Province",
+        "grade": "10,11",
+        "school": "@TLH-1, Faculty of Technological Studies, University of Vavuniya",
+        "date": "25th June 2025",
+        "province": "North Province",
+        "program": "CAREER COMPASS",
+        "comment": "Actually thank you for this program. I like this. Speakers explained the content very well. I really got a lot of information about my higher studies. This is so enjoyable. I think this programme will make me into a better person. I want to choose Biostream. Thank you for everything."
+    },
+    {
+        "id": 5,
+        "name": "Career Compass – Career Guidance Session for the Central Province",
+        "grade": "10,11",
+        "school": "@E.O.E. Pereira Theatre, Faculty of Engineering,  University of Peradeniya",
+        "date": "7th July 2025",
+        "province": "Central Province",
+        "program": "CAREER COMPASS",
+        "comment": "Great work! This was really useful, especially the website. It was nice to know what the future holds. It is great to see the Sri Lanka Inspire community to guide us to the correct path and show what path we can take to decide the future. This was truly useful for us. Thank you!"
+    },
+    {
+        "id": 6,
+        "name": "Career Compass – Career Guidance Session for the Central Province",
+        "grade": "10,11",
+        "school": "@E.O.E. Pereira Theatre, Faculty of Engineering,  University of Peradeniya",
+        "date": "7th July 2025",
+        "province": "Central Province",
+        "program": "CAREER COMPASS",
+        "comment": "පාසල් සිසුන් වන අපට මෙවැනි ඉතා වටිනා වැඩසටහනක් සංවිධානය කිරීම සම්බන්ධයෙන් විශ්වවිද්‍යාලයට IEEE ආයතනයට අප බෙහෙවින් ම ස්තුතිවන්ත වන අතර, O/L අවසාන කර සිටින අපට A/L  විභාගය පිළිබඳව හා ඉන්පසු තෝරා ගත  යුතු ක්ෂේත්‍ර පිළිබද පුළුල් දැනුමක් ලැබුණි. ඉදිරියටත් මෙවැනි වැඩසටහන් කිරීමට ශක්තිය ධෛර්යය ලැබේවායි. ඔබ සැමට සුභපතමි!"
+    }
+]
